refactor(debug): trim admin-only dashboard state
