+++ tests/client/debug/dashboard.svelte.test.ts
@@ -16,7 +16,6 @@ describe('dashboard.svelte', () => {
     expect(state.connected).toBe(false)
     expect(state.sessions).toBeInstanceOf(Map)
     expect(state.wizards).toBeInstanceOf(Map)
-    expect(state.identityMappings).toBeInstanceOf(Map)
     expect(state.activeConfigEditors).toBeInstanceOf(Set)
     expect(state.logScopes).toBeInstanceOf(Set)
     expect(state.logs).toEqual([])
@@ -24,19 +23,11 @@ describe('dashboard.svelte', () => {
     expect(state.turns).toEqual([])
     expect(state.notifications).toEqual([])
     expect(state.toolFailures).toEqual([])
-    expect(state.recurringTasks).toEqual([])
-    expect(state.deferredPrompts).toEqual([])
-    expect(state.memos).toEqual([])
-    expect(state.authorizedGroups).toEqual([])
     expect(state.activeContext).toBe('all')
     expect(state.activeLogFilter).toEqual({})
     expect(state.stats.totalMessages).toBe(0)
     expect(state.stats.totalLlmCalls).toBe(0)
     expect(state.stats.totalToolCalls).toBe(0)
-    expect(state.billingWindow).toBe('30d')
-    expect(state.billingSubjects).toEqual([])
-    expect(state.billingDetail).toBeNull()
-    expect(state.adminLlm).toBeNull()
   })
 
   test('state mutations are observable on the same reference', () => {
